4-Webcam-Photo-Sharer/requirements.txt
wheel
chardet
docutils
filestack-python
idna
Kivy
Kivy-Garden
Pygments
requests
trafaret
urllib3